added the controller recipe get list benchamrk for 500 requests
